Added mixin to the Aether. The sky no longer turns black between y=0-63
This fixes issue #12. There is one mixin applied to WorldRenderer to avoid turning the player's horizon black when below y=63. In ClientProxy, there is a new event handler for EntityViewRenderEvent.FogColors. The point of this is to reverse some calculations done in FogRenderer that causes the fog to get darker as the viewer approaches y=0. Unfortunately, since we can't divide by zero, the void will still be black below y=0.