refactor: The slider uses the goal selector now. (#1835)
Due to the complexity of Minecraft's brains, it's difficult to learn, and this is not good for the long-term maintainability of any modded entity that uses them. By switching the slider AI to goals, hopefully it will become easier to maintain as active team members change, and hopefully gain a slight performance boost.